refactor: introduce TypeVarEmitter.bridgeToKind for target-TypeVar dispatch (Phase 5)
Mirror of `bridgeFromKind` for the symmetric direction (concrete
source, TypeVar target). Canonical contract: `result.representation
== target` — bytes are converted to the type's default form for the
target's kind, then relabeled.

Migrated drop-in matches:
- SumListEmitterCommon (already always-relabel for all 3 kinds)
- ProdBuiltinArrayEmitter (already always-relabel; arrayRepr =
  defaultRepresentation for BuiltinArray, so the Transparent arm's
  switch from arrayRepr to SirTypeUplcGenerator.defaultRepresentation
  is a no-op)

Sites that historically skip the relabel (SumDispatch.sumCaseImpl
Fixed arm, SumUplcConstrEmitter Fixed arm, ProductCaseSirTypeGenerator
catch-all Fixed arm) stay open-coded — migrating those would change
downstream `.representation` checks and needs case-by-case audit.

The local `docs/local/claude/compiler/sum-prod-dispatch-design.md`
(gitignored) has been updated with the audit findings: why universal
outboundStep coverage doesn't fit, what landed instead (per-emitter
`outboundStep` for fixed-source emitters + `via*` helpers +
bridgeFromKind/bridgeToKind), the inconsistencies surfaced
(target-TypeVar relabel divergence, the Transparent-source
KnightsTest:475 invariant).

9 jvm test suites green (4673 tests).
